Refactoring Moved components and systems to relevant modules
