+++ src/components/enemy.rs
@@ -1,0 +1,6 @@
+use bevy::prelude::{Component, Vec2};
+
+#[derive(Component)]
+pub struct Enemy {
+    pub direction: Vec2,
+}
