--- src/components/entities/player/player.rs
@@ -1,1 +1,0 @@
-
