+++ src/components/player.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component)]
+pub struct Player {}
