+++ src/events/game_over.rs
@@ -1,0 +1,3 @@
+pub struct GameOver {
+    pub score: u32,
+}
