+++ src/resources/score.rs
@@ -1,0 +1,12 @@
+use bevy::prelude::Resource;
+
+#[derive(Resource)]
+pub struct Score {
+    pub value: u32,
+}
+
+impl Default for Score {
+    fn default() -> Score {
+        Score { value: 0 }
+    }
+}
