+++ src/systems/enemy.rs
@@ -1,0 +1,85 @@
+use bevy::{prelude::*, window::PrimaryWindow};
+use rand::random;
+
+use crate::{
+    components::{enemy::Enemy, player::Player},
+    events::game_over::GameOver,
+    resources::score::Score,
+};
+
+use super::player::PLAYER_SIZE;
+
+const ENEMY_SIZE: f32 = 60.0;
+const ENEMY_COUNT: u32 = 10;
+const ENEMY_SPEED: f32 = 100.0;
+
+pub fn spawn_enemies(
+    mut commands: Commands,
+    window_query: Query<&Window, With<PrimaryWindow>>,
+    asset_server: Res<AssetServer>,
+) {
+    let window = window_query.get_single().unwrap();
+
+    for _ in 0..ENEMY_COUNT {
+        let random_x = random::<f32>() * window.width();
+        let random_y = random::<f32>() * window.height();
+
+        commands.spawn((
+            SpriteBundle {
+                sprite: Sprite {
+                    custom_size: Option::Some(Vec2::new(ENEMY_SIZE, ENEMY_SIZE)),
+                    ..default()
+                },
+                transform: Transform::from_xyz(random_x, random_y, 0.0),
+                texture: asset_server.load("sprites/zombie.png"),
+                ..default()
+            },
+            Enemy {
+                direction: Vec2::new(random::<f32>(), random::<f32>()).normalize(),
+            },
+        ));
+    }
+}
+
+pub fn change_enemy_direction(mut enemy_query: Query<(&mut Sprite, &Transform, &Enemy)>) {
+    for (mut sprite, transform, enemy) in enemy_query.iter_mut() {
+        let direction = Vec3::new(enemy.direction.x, enemy.direction.y, 0.0);
+        println!("direction {}", enemy.direction.x);
+    }
+}
+
+pub fn enemy_movement(
+    mut enemy_query: Query<(&mut Transform, &Enemy), Without<Player>>,
+    player_query: Query<&Transform, With<Player>>,
+    time: Res<Time>,
+) {
+    let player_transform = player_query.single();
+
+    for (mut transform, enemy) in enemy_query.iter_mut() {
+        let direction = Vec3::new(0.0, 0.0, 0.0);
+        transform.translation += direction * ENEMY_SPEED * time.delta_seconds();
+    }
+}
+
+pub fn enemy_hit_player(
+    mut commands: Commands,
+    mut game_over_event_writer: EventWriter<GameOver>,
+    mut player_query: Query<(Entity, &Transform), With<Player>>,
+    enemy_query: Query<&Transform, With<Enemy>>,
+    score: Res<Score>,
+) {
+    if let Ok((player_entity, player_transform)) = player_query.get_single_mut() {
+        for enemy_transform in enemy_query.iter() {
+            let distance = player_transform
+                .translation
+                .distance(enemy_transform.translation);
+
+            let player_radius = PLAYER_SIZE / 2.0;
+            let enemy_radius = ENEMY_SIZE / 2.0;
+            if distance < player_radius + enemy_radius {
+                // commands.entity(player_entity).despawn();
+                game_over_event_writer.send(GameOver { score: score.value });
+            }
+        }
+    }
+}
